add infinite scrolling support
Re-factor the asset list loading to support the notion of extending the
current set of assets. Instead of storing the full asset resources,
store a summary and stash the full resources in an assetsByUrl object.
This will allow the asset edit form to retrieve an asset without having
to perform an extra fetch. (Or, similarly, let's it populate the form
while the request is in flight.)

The asset list in the redux store is now viewed as a partial list of
assets which may be extended via the getMoreAssets action. The first
load of the asset list is therefore just a special case of dispatching
getMoreAssets with the asset list being empty.

The AssetList component is now completely pure. The AssetTable component
simply renders the current list of assets and the AssetList makes use of
a new GetMoreAssets component. This component will dispatch the
getMoreAssets action when it becomes visible.

In this way, initial asset list load and subsequent loads when one
reaches the bottom of the list are no-longer separate cases. An empty
asset list will have the GetMoreAssets component visible.