add confirmation of asset deletion
Add new action 'confirmDelete' which is dispatched with an asset URL and
the user is asked to confirm deletion of the asset. Managed via a
deleteConfirmation object in the redux store which stores the URL of the
asset currently being considered for deletion.

Add a DeleteConfirmationDialog component which shows a dialog when there
is an asset being considered for deletion and deletes the asset if the
user provides confirmation.