add an asset "draft" object to the store
Add draft asset and associated actions to the redux store. This is
useful for editing an asset.

Add actions to create a new draft and to update it via patches. An
entire asset can be loaded from the store/API via the loadDraft action
which first checks to see if we have a local copy of the asset in
assetsByUrl and, if not, loads it from the API.