add docker-compose for easier developer bootstrap
Currently is is undocumented that the IAR backend needs to be running on
localhost:8000, the lookupproxy on localhost:8080 and both need to be
configured to use the deployed hydra endpoint.

Add a docker-compose configuration which deploys the latest master
branch from lookupproxy and iar-backend on the correct ports and
configures them to use a local hydra instance.

Add documentation to the README outlining how this is to be used.

With this a developer can stand up a completely independent set of
infrastructure to develop on the IAR without depending on any external
services beyond lookup.